Completed part 2 of assignment 1 (added calculate2() method). Still not writing to file as required. Need to add file writing function.
TODO: fix complement method. there is something wrong in that
For ex: if X is set of all ppl who watched movie A,
!X is the complement of X wrt all ppls. 
but currently (all ppl - ppl who watched A) is not equal to
size of !X. need to fix it!

added setops.java a file needed to provide easy set operations (though
it uses underlying retainall, addall etc set methods, named to make
the ops convenient to use).